Refactor the tests, generalize for approximate measures
Meaning, add a state in which the only way for an approximate test to
fail is returning something too good to be true ie. below the optimal
length.

Also add another, randomly generated point series. Broke a cardinal
rule of testing by using the output of the program to be tested in the
test itself (er, it's a regression test?)